Creates trait and moves shared methods to the trait.
